+++ src/z80asm2/cmdline.hpp
@@ -1,0 +1,14 @@
+//-----------------------------------------------------------------------------
+// Parse command line arguments
+// Copyright (c) Paulo Custodio, 2015-2016
+// License: http://www.perlfoundation.org/artistic_license_2_0
+//-----------------------------------------------------------------------------
+
+#ifndef CMDLINE_HPP
+#define CMDLINE_HPP
+
+#include "global.hpp"
+
+extern bool parse_cmdline(int argc, char *argv[], args_t& out_args);
+
+#endif /* ndef CMDLINE_HPP */