+++ src/z80asm2/messages.cpp
@@ -9,7 +9,9 @@ stlplus::message_handler g_errors(std::cerr, 0);	// infinite error limit
 const char *MES_COPYRIGHT = "MES_COPYRIGHT";
 const char *MES_USAGE = "MES_USAGE";
 const char *MES_HELP = "MES_HELP";
-const char *ERR_CMD_LINE_ARGS = "ERR_CMD_LINE_ARGS";
+const char *ERR_INVALID_OPTION = "ERR_INVALID_OPTION";
+const char *ERR_READ_FILE = "ERR_READ_FILE";
+const char *ERR_READ_CONTEXT = "ERR_READ_CONTEXT";
 
 void init_messages() {
 g_messages.add_message(MES_COPYRIGHT, "Z80 Module Assembler 3.0-alpha, (c) Paulo Custodio 2011-2016");
@@ -19,7 +21,9 @@ g_messages.add_message(MES_HELP, "Using z80asm2:\n"
 "  To assemble 'fred.asm' use 'fred' or 'fred.asm'\n"
 "\n"
 "  <modulefile> contains list of file names of all modules to be linked,\n"
-"  one file or another \\@modulefile per line.\n"
+"  one file per line. \n"
+"  Empty lines or lines starting with ';' or '#' are ignored.\n"
+"  Lines starting with '\\@' are expanded recursively.\n"
 "\n"
 "  File types recognized or created by z80asm:\n"
 "    .asm = source file (default), or alternative -e<ext>\n"
@@ -69,6 +73,8 @@ g_messages.add_message(MES_HELP, "Using z80asm2:\n"
 "  -l, --list             Create listing file.lis\n"
 "  -m, --map              Create address map file.map\n"
 "  -g, --globaldef        Create global definition file.def");
-g_errors.add_message(ERR_CMD_LINE_ARGS, "Invalid command line argumens, run z80asm2 -h for help");
+g_errors.add_message(ERR_INVALID_OPTION, "invalid option '@0', run z80asm2 -h for help");
+g_errors.add_message(ERR_READ_FILE, "cannot read file '@0'");
+g_errors.add_message(ERR_READ_CONTEXT, "while reading file '@0'");
 }